backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.15.2
django-filter==25.2
django-cors-headers==4.4.0
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-decouple==3.8
sqlparse==0.5.4
tzdata==2025.3
